refactor(ipc): extract model commands into commands/models.rs (#82)
The IPC layer's commands.rs has been on a refactor diet — meeting
moved out under PR #168, leaving 1.9k LOC. Models is the next
biggest cohesive group: 5 commands (model_list, model_select,
model_download, model_cancel_download, model_remove), 4 result /
event types (ModelCard, ModelSelectResult, DownloadProgress,
DownloadStatus), and the auto-download orchestration that wraps
`crate::transcription::download::download_with_progress`. ~330
lines moved as one block.

Move:

- New `commands/models.rs` with the model block + necessary
  imports (`crate::transcription::catalog`,
  `crate::transcription::download`, `settings::keys`).
- The `poisoned` helper in `commands/mod.rs` becomes
  `pub(super)` so submodules (meeting and now models) can use it.
- Module declaration `pub mod models;` in `commands/mod.rs`,
  alongside the existing `pub mod meeting;`.
- `lib.rs` `generate_handler!` invocations updated from
  `ipc::commands::model_*` to `ipc::commands::models::model_*`.
  Per learnings.md 2026-04-25 the macro is path-sensitive — see
  the comment on `pub mod meeting;` which now applies to both.
- Drop now-unused imports from `commands/mod.rs`:
  `Emitter`, `Manager`, `settings_keys` alias (the remaining
  call sites use the full path), `ModelMetadata`, `download::*`,
  `CancelHandle`.

`commands/mod.rs` drops from 1913 → 1586 LOC. With meeting +
models out, the remaining "big domain" extraction worth
considering is macOS permissions (~250 LOC, three commands +
two result types). Tracked under #82 still.

### Smoke

- `cargo build`, `cargo clippy --all-targets -- -D warnings` clean
- `cargo test --lib` 214 passed (no test moves; mod.rs's tests
  block stays put; the model commands had no unit tests)
- `npm run tauri dev` launches; model picker still functions
