+++ src-tauri/src/lib.rs
@@ -189,11 +189,11 @@ pub fn run() {
             ipc::commands::vocabulary_create,
             ipc::commands::vocabulary_update,
             ipc::commands::vocabulary_delete,
-            ipc::commands::model_list,
-            ipc::commands::model_select,
-            ipc::commands::model_download,
-            ipc::commands::model_cancel_download,
-            ipc::commands::model_remove,
+            ipc::commands::models::model_list,
+            ipc::commands::models::model_select,
+            ipc::commands::models::model_download,
+            ipc::commands::models::model_cancel_download,
+            ipc::commands::models::model_remove,
             ipc::commands::get_first_run_completed,
             ipc::commands::mark_first_run_completed,
             ipc::commands::reset_first_run,
